feat(cache): expose shared Cargo cache integration
