+++ Cargo.toml
@@ -3,8 +3,10 @@ resolver = "3"
 members = [
   "crates/mbx",
   "crates/mbx-cache-core",
+  "crates/mbx-cache-cargo",
   "crates/mbx-cache-protocol",
   "crates/mbx-cache-rustc",
+  "crates/mbx-cache-store",
 ]
 
 [workspace.package]
