+++ crates/mbx-cache-cargo/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "mbx-cache-cargo"
+version = "0.1.0"
+description = "Unstable Cargo invocation integration for mbx cache embedders"
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+readme.workspace = true
+
+[dependencies]
+dirs = "6"
+mbx-cache-core = { path = "../mbx-cache-core", version = "0.5.1", default-features = false }
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+toml = "1"
+
+[dev-dependencies]
+tempfile = "3"
+
+[features]
+default = ["rustls"]
+native-tls = ["mbx-cache-core/native-tls"]
+rustls = ["mbx-cache-core/rustls"]
+rustls-native-roots = ["mbx-cache-core/rustls-native-roots"]
