+++ crates/mbx-cache-core/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "mbx-cache-core"
 version = "0.5.1"
-description = "mbx internals: CAS, authentication, transport, and the cache agent. No API stability -- use the mbx CLI or mbx-cache-protocol."
+description = "Unstable CAS, transport, and cache-agent primitives for mbx embedders"
 edition.workspace = true
 rust-version.workspace = true
 license.workspace = true
