+++ crates/mbx-cache-store/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "mbx-cache-store"
+version = "0.1.0"
+description = "Unstable shared action-store retention and inspection for mbx embedders"
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+readme.workspace = true
+
+[dependencies]
+eyre = "0.6"
+fslock = "0.2"
+log = "0.4"
+mbx-cache-core = { path = "../mbx-cache-core", version = "0.5.1", default-features = false }
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+tempfile = "3"
+
+[dev-dependencies]
+filetime = "0.2"
+
+[features]
+default = ["rustls"]
+native-tls = ["mbx-cache-core/native-tls"]
+rustls = ["mbx-cache-core/rustls"]
+rustls-native-roots = ["mbx-cache-core/rustls-native-roots"]
